#0: rename common.hpp to common/constants.hpp"
