+++ ttnn/cpp/ttnn/operations/data_movement/concat/concat.hpp
@@ -4,7 +4,7 @@
 
 #pragma once
 
-#include "ttnn/common.hpp"
+#include "ttnn/common/constants.hpp"
 #include "ttnn/tensor/types.hpp"
 #include "ttnn/operations/core/core.hpp"
