+++ ttnn/cpp/ttnn/operations/data_movement/tilize_with_val_padding/tilize_with_val_padding.hpp
@@ -7,7 +7,7 @@
 #include "device/tilize_with_val_padding_op.hpp"
 #include "ttnn/run_operation.hpp"
 #include "ttnn/decorators.hpp"
-#include "ttnn/common.hpp"
+#include "ttnn/common/constants.hpp"
 
 namespace ttnn {
 namespace operations::data_movement {
