+++ ttnn/cpp/ttnn/operations/eltwise/binary_backward/binary_backward.hpp
@@ -5,7 +5,7 @@
 
 #pragma once
 
-#include "ttnn/common.hpp"
+#include "ttnn/common/constants.hpp"
 #include "device/binary_backward_op.hpp"
 #include "ttnn/device_operation.hpp"
 #include "ttnn/operations/data_movement.hpp"
