+++ ttnn/cpp/ttnn/operations/experimental/transformer/device/transformer_device_operation.hpp
@@ -6,7 +6,7 @@
 
 #include <optional>
 
-#include "ttnn/common.hpp"
+#include "ttnn/common/constants.hpp"
 #include "ttnn/tensor/tensor.hpp"
 #include "ttnn/run_operation.hpp"
