+++ ttnn/cpp/ttnn/operations/matmul/matmul.cpp
@@ -4,7 +4,7 @@
 
 #include "matmul.hpp"
 
-#include "ttnn/common.hpp"
+#include "ttnn/common/constants.hpp"
 #include "ttnn/operations/core/core.hpp"
 #include "ttnn/operations/data_movement/transpose/transpose.hpp"
 #include "ttnn/operations/eltwise/binary/binary.hpp"
