+++ ttnn/cpp/ttnn/operations/reduction/argmax/device/argmax_op.hpp
@@ -6,7 +6,7 @@
 
 #include <optional>
 
-#include "ttnn/common.hpp"
+#include "ttnn/common/constants.hpp"
 #include "ttnn/tensor/tensor.hpp"
 #include "ttnn/run_operation.hpp"
