+++ ttnn/cpp/ttnn/operations/reduction/topk/topk.hpp
@@ -4,7 +4,7 @@
 
 #pragma once
 
-#include "ttnn/common.hpp"
+#include "ttnn/common/constants.hpp"
 #include "ttnn/decorators.hpp"
 #include "ttnn/operations/core/core.hpp"
